Use common index file on local file system
I used to think that there was a problem sharing an index file for
backup plans. That there was a problem where files exluded from a plan
gets backed up anyway because they are included in the index (if another
backup plan includes them). But testing this now it seems to work the
way it should. So let's try it. After all, bup was designed to work with
a local and shared index.