Polish the UI for selecting backup destination
Adds information texts under the radio buttons, modifies the DriveSelection
class to keep track of if the selected drive is mounted. Also now keeps track
of if any drive is selected or not. Use these to hide/disable drive destination
folder selection widgets. Adds a new dialog for selecting a folder on the external
drive and a tool button for showing the dialog.
Also rewrite and simplify the DriveSelection class, makes it look better in a few
rare cases... and the code easier to read and modify.

And some renaming of method names and cleanup of indentation.. bonus:)