Remove compression level option
It required using an unreleased version of bup, but more importantly it would mostly be useful for developers with tons of source code (what other files could you have that are not already a compressed file format?) that for some reason they wanted to backup. And at the same time willing to use more cpu for the sake of maybe saving a bit of space. Too unlikely. Remove.