Fix opening of symlinks in bup kioslave.
Learning some more about kio... It does not copy symlinks, it just takes the target info from the stat() and then creates a new symlink on destination using that. But it currently only does that if src and dst are the same kioslave. Nevermind that for now, not a big problem. But it anyway, the intention is that calling get() on a symlink should download the target file. So we do that now.