Make libgit2 close packfile file descriptors when done.
A problem is that bup archives can easily have over 1000 pack files, which is not normal for a git repo. libgit2 has not been designed for that, it keeps file descriptors open (presumably as an optimization).
This is a bit hackish fix since it's only done in one place. Seems to be enough for now.