Try to reduce confusion about what will get backed up
The previous logic was to show a dashed (partial) checkmark for folders that has has something inside it that will get backed up, but the whole folder will not be included.
That caused some confusion, for example: if a folder "top" contains three subfolders A, B and C, then the following two scenarios would look the same in the selection widget but actually be very different:
1. "top" selected, "A" and "B" deselected
2. "C" selected
In both cases the "top" folder would be partially checked.

This change tries to make it clearer by only showing "top" folder partially checked in scenario 1. In scenario 2 it will not have a checkmark at all but will still be rendered with "active" text color, not the "disabled" color.
Also adds two more tool tip texts to clarify in more detail why it is shown the way it is.