Add filescanning to the KCM.
Fixes issue #39. Allow user to include (and as a bug fix, also exclude)
folders which the user does not have read permission for. Add filescanning
in the background to detect access permission problems for both files and
folders, then show error message to user about the problem.
This is not a perfect solution, but the best compromise I could think of.
With this solution error message is only displayed during
configuration.
The main problem with that is if unreadable files are added after the
backup plan is created then taking a new backup will fail but no simple
error message.