Add detection of symlink pointing to outside of backup-set.
Fixes issue #38. Since there now is filescanning in the settings
module, use that to detect potential problems and warn the user about it.
The problem is of course again that such problematic symlink could be
added by the user after having created the backup plan. Again, it's a
trade off, don't want to do more filescanning than necessary.