Don't show tray icon when destination available unless manual plan
Based on user feedback, the icon is not so important to have available
easily. For the normal scenario, plugging in a hard drive and maybe taking
a new backup if it's time, the icon is not needed. It is needed in order
to show files for restoring (not common), show log file (not common),
trigger manual backup (mostly interesting if automatic scheduling is off,
which is why that the icon is still shown in that case), or for opening
configuration module (not common).

Also remove the animation that was shown during saving. Not needed anymore
after progress reporting via kuiserver was added.